[refactor] Split helpers/file.ts along its actual seams
file.ts conflated four unrelated concerns in one 670+ line module:
generic fs primitives, archive extraction/creation (incl. filesMove's
package-format sorting), OS-default install-directory resolution, and
privileged installer execution/admin elevation. Split into fs.ts,
archive.ts, paths.ts, and installer.ts along those seams.

file.ts itself becomes a pure re-export barrel (`export * from
'./fs.js'` etc.), so every existing import of '../helpers/file.js'
- across ManagerLocal.ts's remaining callers, all test files, and the
package's public index.ts/index-browser.ts exports - keeps working
unchanged with zero behavior difference. Updated the primary internal
consumers (ManagerLocal.ts, ConfigLocal.ts, RegistryLocal.ts, admin.ts,
packageLocal.ts) to import from the specific new module they actually
need instead of the barrel, since that's the actual point of the split
- a reader can now tell what a file touches (generic fs vs. archive
handling vs. privileged execution) from its imports alone.

NOTE: local verification of this branch is currently blocked - see
conversation for details on an unrelated live-registry-data-drift issue
being fixed on other branches first. `npx tsc --noEmit` and `eslint .`
both pass cleanly; full `npm test` has not yet been re-verified after
rebasing onto the updated main.
